Extract post api
But I'm not sure what this buys me -- I cannot use require.context in here because the version of `require` is different, even during pre-rendering of getStaticProps. My goal was to be able to cache this and call it for each page to generate next and previous links, but I can't do that because MDX doesn't allow me to use getStaticProps.